Move normalization into the wake provider
