+++ tests/components/assist_pipeline/snapshots/test_init.ambr
@@ -294,6 +294,7 @@
         'wake_word_output': dict({
           'timestamp': 2000,
           'wake_word_id': 'test_ww',
+          'wake_word_phrase': 'Test Wake Word',
         }),
       }),
       'type': <PipelineEventType.WAKE_WORD_END: 'wake_word-end'>,
