+++ tests/components/assist_pipeline/snapshots/test_websocket.ambr
@@ -381,6 +381,7 @@
     'wake_word_output': dict({
       'timestamp': 0,
       'wake_word_id': 'test_ww',
+      'wake_word_phrase': 'Test Wake Word',
     }),
   })
 # ---
@@ -966,13 +967,14 @@
     'wake_word_output': dict({
       'timestamp': 0,
       'wake_word_id': 'test_ww',
+      'wake_word_phrase': 'Test Wake Word',
     }),
   })
 # ---
 # name: test_wake_word_cooldown_different_entities.5
   dict({
     'code': 'duplicate_wake_up_detected',
-    'message': 'Duplicate wake-up detected for test ww',
+    'message': 'Duplicate wake-up detected for Test Wake Word',
   })
 # ---
 # name: test_wake_word_cooldown_different_ids
@@ -1026,6 +1028,7 @@
     'wake_word_output': dict({
       'timestamp': 0,
       'wake_word_id': 'test_ww',
+      'wake_word_phrase': 'Test Wake Word',
     }),
   })
 # ---
@@ -1034,6 +1037,7 @@
     'wake_word_output': dict({
       'timestamp': 0,
       'wake_word_id': 'test_ww_2',
+      'wake_word_phrase': 'Test Wake Word 2',
     }),
   })
 # ---
@@ -1088,12 +1092,13 @@
     'wake_word_output': dict({
       'timestamp': 0,
       'wake_word_id': 'test_ww',
+      'wake_word_phrase': 'Test Wake Word',
     }),
   })
 # ---
 # name: test_wake_word_cooldown_same_id.5
   dict({
     'code': 'duplicate_wake_up_detected',
-    'message': 'Duplicate wake-up detected for test ww',
+    'message': 'Duplicate wake-up detected for Test Wake Word',
   })
 # ---
